junior_001/src/main/java/ru/job4j/inout/readconfig/Config.java
package ru.job4j.inout.readconfig;

import java.io.BufferedReader;
import java.io.FileReader;
import java.util.HashMap;
import java.util.Map;
import java.util.StringJoiner;

public class Config {

    private final String path;
    private final Map<String, String> values = new HashMap<>();

    public Config(final String path) {
        this.path = path;
    }

    public void load() {
        String line;
        try (BufferedReader reader = new BufferedReader(new FileReader(this.path))) {
            while ((line = reader.readLine()) != null) {
                if (!line.startsWith("#") && !line.equals("")) {
                    String[] tokens = line.split("=");
                    values.put(tokens[0], tokens[1]);
                }
            }
        } catch (Exception e) {
            e.printStackTrace();
        }
    }

    public String value(String key) {
        return values.get(key);
    }

    @Override
    public String toString() {
        StringJoiner out = new StringJoiner(System.lineSeparator());
        try (BufferedReader reader = new BufferedReader(new FileReader(this.path))) {
            reader.lines().forEach(out::add);
        } catch (Exception e) {
            e.printStackTrace();
        }
        return out.toString();
    }
}
